pyproject.toml
[project]
name = "sub2api-plugin"
version = "0.0.1"
description = "sub2api plugin"
requires-python = ">=3.12"
dependencies = [
    "dify_plugin>=0.4.0,<0.7.0",
    "pytest>=8.0.0,<9.0.0",
    "pyyaml>=6.0.0,<7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
only-include = ["main.py", "manifest.yaml", "provider", "models"]

[tool.pytest.ini_options]
testpaths = ["tests"]
